Introduced Context
Context is a RefCounted JSRuntime and JSContext Impl. it allows for Safe access of Cached Values. if all Values are Gone and the Runtime the Context will free itself.

Value Free Leaks: In cases of Copy Assign Values did not Free itself, also Value uses now the RefCount of the JS Object itself...